requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.118.0
filelock==3.19.1
filterpy==1.4.5
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
kiwisolver==1.4.9
lap==0.5.12
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
polars==1.33.1
psutil==7.1.0
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.25.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
thop==0.1.1.post2209072238
tifffile==2025.5.10
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.204
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.37.0
